+++ src/app/globals.css
@@ -8,10 +8,10 @@ body {
   padding: 0;
 }
 
-* {
+/* * {
   margin: 0;
   box-sizing: border-box;
-}
+} */
 
 @theme {
   /* 🎨 Кольори */
