+++ src/app/signin/page.tsx
@@ -104,7 +104,7 @@ export default function SignInPage() {
         </div>
 
         <p className="text-center text-sm text-gray-600">
-          Don't have an account?{" "}
+          {"Don't have an account? "}
           <Link href="/signup" className="text-blue-600 hover:underline">
             Sign Up
           </Link>
